Use Boost_INCLUDE_DIRS, not Boost_INCLUDE_DIR
The Boost_INCLUDE_DIR is meant to be passed to FindBoost.cmake as
a hint.  However, it will be left unset if cmake finds Boost via
the boost_xxx-config.cmake scripts in CMake's prefix path.

The Boost_INCLUDE_DIRS variable is the one that FindBoost.cmake
sets for use in other cmake scripts.